Rename feature flag and feature folder
