+++ config/development.json
@@ -12,9 +12,9 @@
 		"mobile-app-banner": true,
 		"navigation": true,
 		"onboarding": true,
+		"payment-gateway-suggestions": true,
 		"remote-inbox-notifications": true,
 		"remote-extensions-list": true,
-		"remote-payment-methods": true,
 		"settings": false,
 		"shipping-label-banner": true,
 		"store-alerts": true,
