+++ src/Features/PaymentGatewaySuggestions/DataSourcePoller.php
@@ -3,7 +3,7 @@
  * Handles polling and storage of specs
  */
 
-namespace Automattic\WooCommerce\Admin\Features\RemotePaymentMethods;
+namespace Automattic\WooCommerce\Admin\Features\PaymentGatewaySuggestions;
 
 defined( 'ABSPATH' ) || exit;
